Do not error if we are unable to send email.
Forgetting to set the environment variable causes the email server to
crash and when it is brought up, whatever action we were attempting
to perform before the crash will be retried. So if for example, we were
trying to create a user before the crash, when the email server comes
back, another attempt will be made to create the user agin leading to
non unique email error. Since all these happen before returning to user,
user will see that user will email uniqueness error even though the user
had been created successfully on the first try.