Moved graphql api to eb_data app.
This is because tests for the api were actually proxy test for the
business logic and it makes sense that they should be in same app
as the business logics they test. Also, when when command is run to
get coverage stats, nothing shows up for those tests because elixir can
not link codes inside one umbrella app to tests in another umbrella app.